requirements.txt
flask
flask-cors
opencv-python
scikit-image
Pillow
reportlab
numpy
customtkinter
